Add jshint as a violations reporter.
Because jshint is (at the moment) not available through a python
library/port, confirming that it is available at runtime needs to
happen through a subprocess call, rather than a python import. The base class
for the violations reporter now has a fallback option of checking for a given
reporter through the CLI; however, the implementing class must provide
the command that verifies it's installed. That command must exit 0 if the
tool is successfully installed. (For jshint, it is `jshint --version`).