Ensure case is normalized on Windows when comparing paths from XML coverage report
This fixes a subtle problem on Windows, which is case-preserving but case-insensitive.
Depending from where the user executed the coverage command to produce the XML file,
the `<sources>` tag might contain a different case in the drive letter: `w:\foo\bar`
vs `W:\foo\bar` (note the case of the first letter). On Windows both paths should
be considered the same.

This patch calls the `os.path.normcase` function, which changes all paths
to lower-case on Windows and is a noop on Linux.

I experienced this problem myself and with this patch this fixes the issue
for me.