Update HTML fixtures for pygment 2.12.0
pygment 2.12.0 slightly changes HTML output when `linenos=table`
is used:

"When ``linenos=table`` is used, the ``<table>`` itself is now
wrapped with a ``<div class="highlight">`` tag instead of placing
it inside the ``<td class="code">`` cell (#632.) With this change,
the output matches the documented behavior."

That affects several of the HTML fixtures used in the snippets
tests. This updates them to match pygment's new format. Of
course, this means the tests now only pass with pygment 2.12.0
or higher.

Signed-off-by: Adam Williamson <awilliam@redhat.com>